Move ServiceAdd/RemoveInstanceError
Move the ServiceAdd- and ServiceRemoveInstanceError error enums to the
service::network::error module, and export them from the
service::network module.
